# app/static/swagger.json
{
  "definitions": {
    "PostNew": {
      "properties": {
        "authorId": {
          "type": "integer"
        },
        "content": {
          "type": "string"
        },
        "teamId": {
          "type": "integer"
        }
      },
      "required": [
        "authorId",
        "content"
      ],
      "type": "object"
    },
    "TournamentNew": {
      "properties": {
        "game": {
          "type": "string"
        },
        "maxParticipants": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "startDate": {
          "format": "date-time",
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      },
      "required": [
        "name",
        "game",
        "status",
        "maxParticipants",
        "startDate"
      ],
      "type": "object"
    }
  },
  "info": {
    "description": "A RESTful API for a sports team and player matching platform.",
    "title": "Sport Matching API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticates a user and returns access and refresh tokens.<br/>",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "email": {
                    "example": "user@example.com",
                    "type": "string"
                  },
                  "password": {
                    "example": "password123",
                    "format": "password",
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Authentication successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "User Login",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Invalidates the user's refresh token.<br/>",
        "responses": {
          "200": {
            "description": "Successfully logged out."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Logout user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Uses a valid refresh token to get a new access token.<br/>",
        "responses": {
          "200": {
            "description": "Access token refreshed successfully."
          },
          "401": {
            "description": "Invalid or expired refresh token."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "description": "Creates a new user and a corresponding profile based on the specified role.<br/>",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "city": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "nickname": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "role": {
                    "enum": [
                      "player",
                      "referee",
                      "coach"
                    ],
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password",
                  "nickname",
                  "role"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "User created successfully."
          },
          "400": {
            "description": "Missing required fields."
          },
          "409": {
            "description": "User with this email or nickname already exists."
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/v1/posts": {
      "get": {
        "responses": {
          "200": {
            "description": "A list of posts.",
            "schema": {
              "items": {
                "$ref": "#/definitions/Post"
              },
              "type": "array"
            }
          }
        },
        "summary": "Retrieve all posts.",
        "tags": [
          "Legacy"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PostNew"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Post created successfully"
          },
          "400": {
            "description": "Missing data or author/team not found"
          }
        },
        "summary": "Create a new post.",
        "tags": [
          "Legacy"
        ]
      }
    },
...
    "/api/v1/sports": {
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "properties": {
                "isTeamSport": {
                  "type": "boolean"
                },
                "name": {
                  "type": "string"
                }
              },
              "required": [
                "name",
                "isTeamSport"
              ],
              "type": "object"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Sport created successfully"
          },
          "400": {
            "description": "Missing name or isTeamSport, or sport with this name already exists"
          }
        },
        "summary": "Create a new sport",
        "tags": [
          "Sports"
        ]
      }
    },
...
    "/api/v1/tournaments": {
      "get": {
        "responses": {
          "200": {
            "description": "A list of tournaments.",
            "schema": {
              "items": {
                "$ref": "#/definitions/Tournament"
              },
              "type": "array"
            }
          }
        },
        "summary": "Retrieve all tournaments.",
        "tags": [
          "Legacy"
        ]
      },
      "post": {
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/TournamentNew"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Tournament created successfully"
          },
          "400": {
            "description": "Missing data"
          }
        },
        "summary": "Create a new tournament.",
        "tags": [
          "Legacy"
        ]
      }
    },
    "/api/v1/tournaments/{tournament_id}": {
      "get": {
        "parameters": [
          {
            "description": "ID of the tournament to retrieve",
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Tournament found",
            "schema": {
              "$ref": "#/definitions/Tournament"
            }
          },
          "404": {
            "description": "Tournament not found"
          }
        },
        "summary": "Retrieve a single tournament by its ID.",
        "tags": [
          "Legacy"
        ]
      }
    },
...
  },
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
      "in": "header",
      "name": "Authorization",
      "type": "apiKey"
    }
  },
  "swagger": "2.0"
}